+++ vibrato/src/tokenizer/lattice.rs
@@ -14,8 +14,8 @@ const INVALID_IDX: u16 = u16::MAX;
 pub struct Node {
     pub word_id: u32,
     pub lex_type: LexType, // 8 bits
-    pub start_node: u16,
-    pub start_word: u16,
+    pub start_node: usize,
+    pub start_word: usize,
     pub left_id: u16,
     pub right_id: u16,
     pub min_idx: u16,
@@ -39,24 +39,24 @@ impl Node {
 pub struct Lattice {
     ends: Vec<Vec<Node>>,
     eos: Option<Node>,
-    len_char: u16, // needed for avoiding to free ends
+    len_char: usize, // needed for avoiding to free ends
 }
 
 impl Lattice {
-    pub fn reset(&mut self, len_char: u16) {
+    pub fn reset(&mut self, len_char: usize) {
         Self::reset_vec(&mut self.ends, len_char + 1);
         self.len_char = len_char;
         self.eos = None;
         self.insert_bos();
     }
 
-    fn reset_vec<T>(data: &mut Vec<Vec<T>>, new_len: u16) {
+    fn reset_vec<T>(data: &mut Vec<Vec<T>>, new_len: usize) {
         for v in data.iter_mut() {
             v.clear();
         }
-        let cur_len = data.len() as u16;
+        let cur_len = data.len();
         if cur_len <= new_len {
-            data.reserve(usize::from(new_len - cur_len));
+            data.reserve(new_len - cur_len);
             for _ in cur_len..new_len {
                 data.push(Vec::with_capacity(16))
             }
@@ -65,7 +65,7 @@ impl Lattice {
 
     /// Returns the number of characters of the set sentence.
     #[inline(always)]
-    pub const fn len_char(&self) -> u16 {
+    pub const fn len_char(&self) -> usize {
         self.len_char
     }
 
@@ -82,7 +82,7 @@ impl Lattice {
         });
     }
 
-    pub fn insert_eos<C>(&mut self, start_node: u16, connector: &C)
+    pub fn insert_eos<C>(&mut self, start_node: usize, connector: &C)
     where
         C: ConnectorCost,
     {
@@ -100,7 +100,7 @@ impl Lattice {
         });
     }
 
-    pub unsafe fn insert_eos_unchecked<C>(&mut self, start_node: u16, connector: &C)
+    pub unsafe fn insert_eos_unchecked<C>(&mut self, start_node: usize, connector: &C)
     where
         C: ConnectorCost,
     {
@@ -120,9 +120,9 @@ impl Lattice {
 
     pub fn insert_node<C>(
         &mut self,
-        start_node: u16,
-        start_word: u16,
-        end_word: u16,
+        start_node: usize,
+        start_word: usize,
+        end_word: usize,
         word_idx: WordIdx,
         word_param: WordParam,
         connector: &C,
@@ -132,7 +132,7 @@ impl Lattice {
         debug_assert!(start_node <= start_word);
         debug_assert!(start_word < end_word);
         let (min_idx, min_cost) = self.search_min_node(start_node, word_param.left_id, connector);
-        self.ends[usize::from(end_word)].push(Node {
+        self.ends[end_word].push(Node {
             word_id: word_idx.word_id,
             lex_type: word_idx.lex_type,
             start_node,
@@ -146,9 +146,9 @@ impl Lattice {
 
     pub unsafe fn insert_node_unchecked<C>(
         &mut self,
-        start_node: u16,
-        start_word: u16,
-        end_word: u16,
+        start_node: usize,
+        start_word: usize,
+        end_word: usize,
         word_idx: WordIdx,
         word_param: WordParam,
         connector: &C,
@@ -159,7 +159,7 @@ impl Lattice {
         debug_assert!(start_word < end_word);
         let (min_idx, min_cost) =
             self.search_min_node_unchecked(start_node, word_param.left_id, connector);
-        self.ends[usize::from(end_word)].push(Node {
+        self.ends[end_word].push(Node {
             word_id: word_idx.word_id,
             lex_type: word_idx.lex_type,
             start_node,
@@ -171,15 +171,15 @@ impl Lattice {
         });
     }
 
-    fn search_min_node<C>(&self, start_node: u16, left_id: u16, connector: &C) -> (u16, i32)
+    fn search_min_node<C>(&self, start_node: usize, left_id: u16, connector: &C) -> (u16, i32)
     where
         C: ConnectorCost,
     {
-        debug_assert!(!self.ends[usize::from(start_node)].is_empty());
+        debug_assert!(!self.ends[start_node].is_empty());
 
         let mut min_idx = INVALID_IDX;
         let mut min_cost = MAX_COST;
-        for (i, left_node) in self.ends[usize::from(start_node)].iter().enumerate() {
+        for (i, left_node) in self.ends[start_node].iter().enumerate() {
             debug_assert!(left_node.is_connected_to_bos());
             let conn_cost = connector.cost(left_node.right_id, left_id);
             let new_cost = left_node.min_cost + conn_cost;
@@ -197,18 +197,18 @@ impl Lattice {
 
     unsafe fn search_min_node_unchecked<C>(
         &self,
-        start_node: u16,
+        start_node: usize,
         left_id: u16,
         connector: &C,
     ) -> (u16, i32)
     where
         C: ConnectorCost,
     {
-        debug_assert!(!self.ends[usize::from(start_node)].is_empty());
+        debug_assert!(!self.ends[start_node].is_empty());
 
         let mut min_idx = INVALID_IDX;
         let mut min_cost = MAX_COST;
-        for (i, left_node) in self.ends[usize::from(start_node)].iter().enumerate() {
+        for (i, left_node) in self.ends[start_node].iter().enumerate() {
             debug_assert!(left_node.is_connected_to_bos());
             let conn_cost = connector.cost_unchecked(left_node.right_id, left_id);
             let new_cost = left_node.min_cost + conn_cost;
@@ -226,19 +226,16 @@ impl Lattice {
 
     /// Checks if there exist at least one at the word end boundary
     #[inline(always)]
-    pub fn has_previous_node(&self, i: u16) -> bool {
-        self.ends
-            .get(usize::from(i))
-            .map(|d| !d.is_empty())
-            .unwrap_or(false)
+    pub fn has_previous_node(&self, i: usize) -> bool {
+        self.ends.get(i).map(|d| !d.is_empty()).unwrap_or(false)
     }
 
-    pub fn append_top_nodes(&self, top_nodes: &mut Vec<(u16, Node)>) {
+    pub fn append_top_nodes(&self, top_nodes: &mut Vec<(usize, Node)>) {
         let eos = self.eos.as_ref().unwrap();
         let mut end_node = eos.start_node;
         let mut min_idx = eos.min_idx;
         while end_node != 0 {
-            let node = &self.ends[usize::from(end_node)][usize::from(min_idx)];
+            let node = &self.ends[end_node][usize::from(min_idx)];
             top_nodes.push((end_node, node.clone()));
             (end_node, min_idx) = (node.start_node, node.min_idx);
         }
@@ -246,15 +243,15 @@ impl Lattice {
 
     pub fn add_connid_counts(&self, counter: &mut ConnIdCounter) {
         for end_char in 1..=self.len_char() {
-            for r_node in &self.ends[usize::from(end_char)] {
+            for r_node in &self.ends[end_char] {
                 let start_node = r_node.start_node;
-                for l_node in &self.ends[usize::from(start_node)] {
+                for l_node in &self.ends[start_node] {
                     counter.add(r_node.left_id, l_node.right_id, 1);
                 }
             }
         }
         let r_node = self.eos.as_ref().unwrap();
-        for l_node in &self.ends[usize::from(self.len_char())] {
+        for l_node in &self.ends[self.len_char()] {
             counter.add(r_node.left_id, l_node.right_id, 1);
         }
     }
@@ -263,10 +260,7 @@ impl Lattice {
 impl std::fmt::Debug for Lattice {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         writeln!(f, "Lattice {{ eos: {:?}, ends: [", &self.eos)?;
-        for (i, e) in self.ends[..=usize::from(self.len_char())]
-            .iter()
-            .enumerate()
-        {
+        for (i, e) in self.ends[..=self.len_char()].iter().enumerate() {
             writeln!(f, "{} => {:?}", i, e)?;
         }
         writeln!(f, "]}}")
